Fixed getAverageById and added test
I had a bug in getAverageById() where it would return an array of duplicate movies and average ratings. Created private method getUniqueMovieIds() that returns an array of unique movie IDs that is used by getAverageById(). Added code to printAverageRatings() in MovieRunnerAverage so the array of average movie ratings is sorted and each array element is printed once per line. Added test class for MovieRunnerAverage.