Began work on Step Three Programming Assignment One
I added the instructor-provided classes to the coursera.capstone.project package so tests can import them when necessary. I converted Rater to an interface, which required renaming of the original Rater class to PlainRater along with all the calls to it. Created EfficientRater class that implements Rater interface and changes the myRatings field from an ArrayList to a HashMap.